requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.19
langgraph-checkpoint>=0.2.0
langgraph-checkpoint-sqlite>=0.2.0
langchain-community>=0.3.0
python-dotenv>=1.0.1
rich>=13.8.0
ddgs>=9.0.0
sentence-transformers>=3.0.1
transformers>=4.50.0
chromadb>=0.5.0
langfuse>=2.40.2
aiosqlite>=0.20.0
gradio>=4.0.0
